fixup! feat: add GET_PROTOCOL_PARAMETERS
